Started testing.
