+++ tests/phpunit.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<phpunit>
+<phpunit colors="true">
 	<testsuites>
 		<testsuite name="asgrim/ofxparser">
 			<directory>./OfxParser/</directory>
